refactor(server): extract DB + conversion operations from analysis.ts into analysis-db.ts
Separates SQLite persistence (saveInsightsToDb, deleteSessionInsights, saveFacetsToDb)
and data conversion (convertToInsightRows, convertPromptQualityToInsightRow) from LLM
orchestration logic. analysis.ts now focuses on LLM calls, chunking, and response merging.

InsightRow and SessionData types moved to analysis-db.ts (owned by DB layer);
re-exported from analysis.ts for backward compat.
